==> 2-ImageResize.ipynb <==
import os
import pandas as pd
import numpy as np
import matplotlib
import matplotlib.image as mpimg
import matplotlib.pyplot as plt
import matplotlib.gridspec as gridspec

from PIL import Image, ImageChops, ImageOps

def resize_image(f_in, f_out, scale = 0.04):

    image = Image.open(f_in)
    new_size = int(scale * image.size[0]), int(scale * image.size[1])
    resized_image = image.resize(new_size, Image.ANTIALIAS)
    resized_image.save(f_out)

# loadin revised images
image_dir = os.getcwd() + '\\Data\\images\\'
image_dir_src = image_dir + 'rev\\' 
image_dir_dest = image_dir + 'res_rev_04\\' 

# total count of samples
sample_count = 1584

# load images and then resize/save to res_rev 
images = []
for image_label in range(sample_count):
    img_src = image_dir_src + str(image_label + 1) + '.jpg'
    img_dest = image_dir_dest + str(image_label + 1) + '.jpg'
    resize_image(img_src, img_dest)
