==> 3-LabelCleanup.ipynb <==
import os
import time
import pandas as pd
import numpy as np

# load attributes 
data_dir = os.getcwd() + '\\Data\\'
train_data = pd.read_csv(data_dir + 'train.csv')
data = train_data.ix[:, ['id', 'species']]
data.head()

labels = list(set(data['species']))
labels = pd.DataFrame(labels)
labels[1] = labels.index
labels.columns = ['species', 'specy_ids']
labels.head(5)

data = pd.merge(data, labels, how='left', on='species')
data.head(5)

data.to_csv(data_dir + 'train_labels.csv', header='True', index=False)

label_count = len(labels)
train_count = len(data)
train_label_bits = np.zeros(shape=(train_count,label_count))
for i in range(train_count):
    train_label_bits[i][data['specy_ids'][i]] = 1

np.savetxt(data_dir + 'train_label_bits.csv', train_label_bits, delimiter=',', fmt='%i')
